feat: initial policy nft contract (#2)
**Motivation:**

Building the initial version of the policy NFT so we can create a
prototype of vertex as soon as possible. The Policy NFT allows for users
to receive soulbound tokens which represent voting power and access
level. Tokens can be given roles, and roles have permissions. Depending
on a tokens permissions, it will allow the token holder to propose
various governance actions to the Vertex Administrator contract.

**Modifications:**

Built the initial version of the `PolicyERC721.sol` contract, added a
`slither.config.json` file, and renamed the `tests` dir to `test` (for
this last one, let me know if there's a way to reconfigure foundry to
find tests in a directory named anything other than `test`.

**Result:**

Working prototype of the Policy NFT to integrate with the rest of the
vertex system