refactor: Emit cleaner events for better off-chain indexing (#40)
**Motivation:**
Improving the following events for better parsing by off-chain indexers:
1. Currently the strategy authorization/unauthorization function emits a
batched event for a list of strategies to be authorized/unauthorized.
Refactor it to emit independent events for each strategy
authorization/unauthorization (with strategy address included).
2. Currently the factory contract emitting a `VertexCreated` event does
not include the policy address in the event. Refactor the event to
include it.

**Modifications:**

1. Emitting independent events for each strategy
authorization/unauthorization (with strategy address included)
2. Uniform salt usage for all Strategy creation
3. Refactor `VertexCreated` event to include Policy address.
4. Fork testing for all tests in the repo
5. Fixing CREATE2 machine dependent bytecode hash issue

**Result:**

Cleaner events for off-chain infrastructure to parse.