test: refactors VertexFactory tests to use one test contract per method (#69)
**Motivation:**

This is the first part of
https://github.com/llama-community/vertex-v1/issues/60, so the primary
goal of this PR is to restructure the tests as described there.

**Modifications:**

The primary changes to this PR are the refactoring of the tests to use
contracts named `Constructor`, `Deploy`, and `ComputeAddress`. These
represent the method names in the `VertexFactory` contract, and the test
contracts mirror the order of the methods in the src contract. The first
two (`Constructor` and `Deploy`) are complete, and the test names should
describe the full spec for those methods. The final one is a placeholder
since the compute address methods don't exist. If you think are tests
are missing (i.e. there is a functionality/behavior/side-effect not
explicitly called out in a test name) leave a comment on this PR

Additional changes are some cleanup to make things more readable:
- I changed a few variable names to better reflect their usage. For
example, `initialVertex` sounds unimportant (why does it matter what the
first vertex deployed is?) but `rootVertex` more clearly reflects the
role (it has root access to the protocol). Similarly, the use of
`vertexCoreLogic` makes it clear that a contract is intended to be used
as the logic contract for a proxy.
- I also took the deploy logic, pulled it into an internal `_deploy`
method, and reused that in both the constructor and the external
`deploy` method
- Function visibilities were fixed: methods were `public` when they
could have more restrictive (`internal` or `external`) visibilities
- In tests, I removed the leading underscores from helper methods. Tests
already call tons of internal methods that don't have leading
underscores (i.e. everything in forge-std) so IMO there's no reason to
use the leading underscore convention in tests, as it just adds noise.
In `src` contracts on the other hand this is a useful convention to make
it explicit when you're calling internal helper methods

**Result:**

Everything described in the modifications section 😅