style: reduce line length to improve readability (#108)
**Motivation:**

Reduce line length from 160 chars to 120 to improve readability.
Additionally updates the directory structure to match what's specified
in https://github.com/llama-community/vertex-v1/issues/104

**Modifications:**

Updated configs to 120 and ran `forge fmt`.

The `src/interfaces` directory has a file called `VertexPolicy.sol`.
This inherits from ERC721 so has to be an abstract contract, which is
why I did not changes this to an `interface`. #104 suggests removing
this file altogether as part of the renames, though, so this is ok for
now since it will get changed later

**Result:**

Solidity and repository will be more readable.

---------

Co-authored-by: Austin Green <austinzgreen@gmail.com>