refactor: remove action queue, rename permissionSignature, and refactor foundry config (#126)
**Motivation:**

This PR includes a bunch of refactors to improve developer experience
and ensure consistency across the codebase.

- Closes #64
- Closes #119
- Closes #121
- Closes #122
- Closes #123

**Modifications:**

It refactors the foundry.toml so running forge commands by default will
work for local development (via-ir and optimizer are set to false and
contract size warning is disabled). This means we can develop in peace
but still ensure the codebase is built and tested properly in CI. It
also updates the README to reflect our switch from a Makefile to
justfile. I renamed `permissionSignature` to `permissionId` across the
codebase, removed the queuedActions mapping (it's being used anywhere),
changed private functions to internal to ensure consistency, removes the
unused `calculateExpiration` function, and moves ActionState to a
dedicated Enums file.

**Result:**

We will have an improved developer experience and more consistent repo.