fix: remove revoke expired permissions (#146)
**Motivation:**

the `revokeExpiredPermissions` function was not serving any purpose, as
it would only be useful if we were enumerating permissions onchain. We
have removed it to reduce the bytecode size of the policy contract and
cleanup unnecessary methods. This Closes #92

**Modifications:**

- Removed `revokeExpiredPermissions`

- Fixed logic in updating permissions expirations, allowing permissions
with an expiration to be set to 0, effectively removing the expiration

- added an empty test to be filled in when Matt's next PR merges which
changes the policy contract significantly

**Result:**

reduced bytecode size in the policy contract, slightly better dev-x with
the `batchUpdatePermissions` function, and no useless functions in the
policy contract.