feat: implement action guards to add checks  (#225)
Closes https://github.com/llama-community/vertex-v1/issues/221,
implements part of #214. #221 is straightforward and explained in that
issue, so the rest of this PR description will cover account guards

**Motivation:**

Imagine you have a vertex account that holds funds for a grant
committee. It holds a lot of funds and some grants will be large grants
and some grants will be small grants. Depending on the grant size you
want different approval thresholds, i.e. different strategies, for how
to approve it. Similarly, you may want to check the USD value of the
grant if it's denominated in tokens. Currently, there is no easy way to
do that but action guards will make this very simple to do.

Guards are keyed off the target/selector combination for a given action.

**Modifications:**

- Add an `IActionGuard` interface that all action guards must conform
to.
- Add a `setGuard` method to vertex core than can be used to add/remove
action guards
- Add checks to `createAction` and `executeAction` to look if a guard is
present for the target/selector, and if so call it to make sure the
action is valid.

**Result:**

Orgs can easily add additional checks and restrictions to actions.