refactor: move logic into strategies, and fix weight bug (#230)
Closes https://github.com/llama-community/vertex-v1/issues/220, closes
https://github.com/llama-community/vertex-v1/issues/227

**Motivation:**

In #227 we had a bug when fetching weight which needs to be fixed

In #220 we discussed moving logic out of core and into strategies to
make things more flexible, as strategies can be changing with new logic
added, but the core cannot be changed

**Modifications:**

- Move cancelation logic, expiration logic, and disapproval logic into
strategy
- Deduplicate a few custom errors that were similar to simplify the ABI
- Also cleaned up a few tests to remove some unnecessary code + fix an
invariant test bug
- Rename the action `executionTime` field to `minExecutionTime`

**Result:**

The logic for (1) determining if an action can be canceled, (2)
determining if an action is expired, and (3) determining if disapprovals
are supported are all handled by the strategy, therefore they can be
changed.

Relative (percentage) vs. absolute voting thresholds can also be
supported now. For example, relative voting would be supported by
changing the strategy's `isActionPassed` from:

```solidity
function isActionPassed(uint256 actionId) external view returns (bool) {
  Action memory action = vertex.getAction(actionId);
  return action.totalApprovals >= getMinimumAmountNeeded(action.approvalPolicySupply, minApprovalPct);
}
```

to 

```solidity
function isActionPassed(uint256 actionId) external view returns (bool) {
  Action memory action = vertex.getAction(actionId);
  return action.totalApprovals >= ACTION_THRESHOLD;
}
```

where `ACTION_THRESHOLD` is a strategy constant. Now the obvious
question is how does this constant value get set? It must be passed to
the `initialize` method, but with the current interface that wouldn't be
possible. Therefore my suggestion is to, in a separate PR:
1. Create an ` interface IVertexStrategy` contract. This is the
interface that strategies MUST conform to, and contains the set of
functions that `VertexCore` (or other contracts in the protocol) will
call on the strategy.
2. Update types from `VertexStrategy` to `IVertexStrategy`
3. Rename the current `VertexStrategy` contract to something more
descriptive, like `VertexStrategyPercentageThresholds`
4. The last thing is that we'll need to change the `initialize` method
and standardize this in the interface—different strategies will have
different parameters, so we'll need to change some parts to a generic
`bytes` argument that the strategy itself decodes.

This combination of changes will make strategies very flexible, with the
only limitations being that they conform to the interface, and
everything else can be changed.

## TWO ASKS:
1. Let me know if there's anything else I missed that should be moved
into the strategy contract
2. Let me know if that interface refactor is what I should work on next