refactor: strategy interface and more flexible strategies (#239)
Closes https://github.com/llama-community/vertex/issues/233, closes
https://github.com/llama-community/vertex/issues/218, closes
https://github.com/llama-community/vertex/issues/234 and makes policies
even more flexible

**Motivation:**

- Part of the motivation is described in those two issues
- The other part is in the description of
https://github.com/llama-community/vertex/pull/230, basically there's
room to make things much more modular/flexible by swapping out strategy
logic

**Modifications:**

Main changes:
- New `IVertexStrategy` interface defines the required methods a
strategy must implement. These methods are all required by `VertexCore`,
with the exception of the `vertex` and `policy` getter methods, which
are simply required for convenience.
- All src contracts use `IVertexStrategy` as the type
- Because different strategies might have different initialization
parameters, they now are initialized with a generic `bytes` arg and we
added `encodeStrategy` and `encodeStrategies` helper methods to handle
this encoding in tests/scripts.
- Deduped some custom errors and deduped some logic from the
`_castApproval` and `_castDisapproval` methods
- Like guards, strategies now have a required `validateActionCreation`
method. This is so checks like "don't allow the action can be created if
the approval role has zero supply" now live in the strategy, so those
checks can be changed. For example, we might be ok with that sometimes,
e.g. if we only care about force approvals for a strategy.
- In fact, the whole concept of "approval roles" and "force approval
roles" (and disapproval) is not strategy specific. VertexCore does not
know and does not care about such a distinction. Therefore a new
strategy can be introduced with only force approval roles very easily
- Similarly, VertexCore doesn't care about supply of roles, this is also
moved to strategy. Therefore a new strategy to allow or disallow action
creators to vote can also be introduced easily.
- Related to that, this PR modifies the default strategy so action
creators cannot vote. This is done by expanding the scope of the
`isApprovalEnabled` checks to let strategies implement any logic they
want.
- Renames `weights` to `quantities` throughout the codebase.

Smaller things:
- Tests got really noisy due to console.logs in the deploy script, so
added a `print` helper method to not show logs if the chain ID is
forge's default of 31337 (called `print` because `log` is already used
for event names in DSTest)

**Result:**

Same as modifications, basically strategies are now super flexible!

---------

Co-authored-by: dd0sxx <theo@llama.xyz>