feat: add scripts and validation to setGuard (#246)
**Motivation:**

Closes #242. Closes #214. This gives Vertex users a system for
delegatecalling from VertexCore. These scripts can be used to make sure
permissioning more closely represents the underlying action that the
permission represents.

**Modifications:**

Add an authorizedScripts mapping, authorizeScript function, logic to
executeAction that delegatecalls scripts, and a validation check to
guards that ensures the target isn't the core or policy.

**Result:**

Users will be able to extend the functionality of their vertex instance
with scripts.