perf: only save hash of action info to storage to reduce gas cost (#265)
This is the same as #264, but now it's been rebased against main

**Motivation:**

Saving all info to execute an action to storage can be prohibitively
expensive

**Modifications:**

Only save a hash to storage, and use a new `ActionInfo` struct to
represent the hashed data. This felt cleaner then adding 5 new
parameters to all methods, which is why I went this route.

Also combined the `selector` and `data` args into a single parameter to
simplify things

**Result:**

Cheaper gas. PR is into a dedicated branch that updated the snapshots to
avoid the difference being affected by other PRs.
Keeping this as draft for now since I'll probably add a few other small
gas changes to this PR before merging, but def feel free to start
reviewing. See the `.gas-report` file for updated snapshots