feat: check slot0 in LlamaCore delegatecalls (#270)
**Motivation:**

Whenever we're executing arbitrary code in the context of LlamaCore, we
want to ensure that none of the storage in this contract changes in
unexpected ways, as this could let someone who sneaks in a malicious (or
buggy) target to effectively take ownership of this contract. However,
this contract has a lot of storage so it's not practical to check all
slots, especially since some may be expected to change. Therefore we
instead just check slot0, since that slot (1) contains core variables
that should never be changed, and (2) is the first slot so it's the most
likely to be accidentally overwritten with a bad script. The storage
layout of this contract is below, generated with `forge inspect
LlamaCore storage --pretty`:

```
| Variable Name        | Type                                                         | Slot | Offset | Bytes |
|----------------------|--------------------------------------------------------------|------|--------|-------|
| _initialized         | uint8                                                        | 0    | 0      | 1     |
| _initializing        | bool                                                         | 0    | 1      | 1     |
| policy               | contract LlamaPolicy                                         | 0    | 2      | 20    |
| factory              | contract LlamaFactory                                        | 1    | 0      | 20    |
| llamaAccountLogic    | contract LlamaAccount                                        | 2    | 0      | 20    |
| name                 | string                                                       | 3    | 0      | 32    |
| actionsCount         | uint256                                                      | 4    | 0      | 32    |
| actions              | mapping(uint256 => struct Action)                            | 5    | 0      | 32    |
| approvals            | mapping(uint256 => mapping(address => bool))                 | 6    | 0      | 32    |
| disapprovals         | mapping(uint256 => mapping(address => bool))                 | 7    | 0      | 32    |
| authorizedStrategies | mapping(contract ILlamaStrategy => bool)                     | 8    | 0      | 32    |
| authorizedScripts    | mapping(address => bool)                                     | 9    | 0      | 32    |
| nonces               | mapping(address => mapping(bytes4 => uint256))               | 10   | 0      | 32    |
| actionGuard          | mapping(address => mapping(bytes4 => contract IActionGuard)) | 11   | 0      | 32    |
```
**Modifications:**

Add checks on the slot 0 value before and after the delegatecall

**Result:**

Malicious/buggy scripts cannot as easily take ownership of the account