feat: allow action creators to provide an optional description (#269)
**Motivation:**

This provides a lightweight way for all clients to show the same title
and description of actions. This will make it easier for action creators
to communicate what they want to accomplish with their action.

**Modifications:**

Overload the `createAction` and `createActionBySig` functions to add an
optional description string that is emitted in the `ActionCreated`
event.

**Result:**

Llama clients will be able to display descriptions for actions.