fix: improve policy nft metadata (#276)
**Motivation:**

After seeing the policy NFT collection on Opensea, we decided to make a
few changes: change the symbol format to `LL-{NAME IN UPPERCASE WITH
SPACES REPLACED WITH '-'}`, improve descriptions, rename
`policyTokenURI` to `policyMetadata`. We also decided to drop the index
for `id` in the `ActionCreated` event.

**Modifications:**

Used solady to compute the symbol, improved the descriptions, fixed
tests.

**Result:**

Our policy NFT will be ready for audit tomorrow.