fix: use dedicated runner and pin foundry version (#304)
**Motivation:**

This change ensures we have a large enough machine to run our foundry
test suite and we're pinned to a stable version of foundry.

**Modifications:**

Pinning foundry to a stable version to avoid an inconsistent CI
experience as we finish our Spearbit followups. We should remove this
change and fix the resulting test failures after we finish all of the
follow ups. I also update the `ci.yml` file to use our dedicated
enterprise runner group.

**Result:**

We will have an improved developer experience for finishing the audit
follow ups.