feat: execute actions through dedicated executor contract (#298)
**Motivation:**

https://github.com/spearbit-audits/review-llama/issues/15

**Modifications:**

Created the `LlamaExecutor` contract, updated `LlamaCore` to use
`LlamaExecutor` for action execution, replaced all `OnlyLlama` modifiers
to use the executor instead of core. The rest of this PR is fixing the
resulting conflicts that arise from making this change. This includes
but isn't limited to: updating `LlamaPolicyMetadataParams` to use the
executor instead of core in the color and logo mapping, updating the
`_deploy` function in the factory to return both the executor and core,
changing all tests to prank as the executor instead of the core when
simulating actions.

**Result:**

We will have additional protection when instances delegatecall scripts.

---------

Co-authored-by: Matt Solomon <matt@mattsolomon.dev>
Co-authored-by: dd0sxx <contacttheotelonis@gmail.com>