refactor: simplify the `LlamaLens` interface by adding the factory as a constructor arg (#323)
**Motivation:**

This was @mds1's idea to improve the interface of many of the
`LlamaLens` helper functions. Protocol code does not use `LlamaLens` so
it's worth improving the experience by requiring as few arguments as
possible.

**Modifications:**

Added a constructor that accepts a `LlamaFactory` contract. Then it uses
this contract parameter to fetch the account, policy and core logic
contracts. After setting these all as immutable we can update the lens
functions accordingly.

**Result:**

Better developer experience when using the lens contract.