feat: hardcode color and logo for root instance on deploy (#342)
**Motivation:**

Closes https://github.com/llamaxyz/llama/issues/334.

https://github.com/llamaxyz/llama/pull/327 did not include setting the
color and logo for the root instance on deploy of the factory. We chose
this method to avoid a stack too deep error and minimizing the amount of
code that needed to be changed. Before deploy of the factory, the root
instance can edit the color and logo variables in the
`LlamaPolicyMetadataParamRegistry.sol` constructor and after deploy
these values can be updated through the normal `setColor` and `setLogo`
flow.

**Modifications:**

Hardcoded the root instance's color and logo and set them in the
`LlamaPolicyMetadataParamRegistry`'s constructor.

**Result:**

The root instance will have the llama color and logo on deploy of the
factory.