build: remove unused solarray dependency
Foundry tries to infer remappings by default, and for ds-test it
was using `ds-test/=lib/solarray/lib/ds-test/src/`. However, solarray
uses forge-std v1.5.2 which uses an older ds-test version that does not
have the assertNotEq helpers that we rely on. This causes compilation
errors (though seemingly only on linux and not macOS). We no longer
use this dependency, so we can remove it.