feat: add SupplyCheckpoints lib for checkpoing total supplies (#413)
**Motivation:**

We may want to checkpoint total supplies to remove a few edge case bugs, and this requires a new library since it's a different struct than balance checkpoints.

**Modifications:**

Adds `SupplyCheckpoints.sol` library which is similar to `Checkpoints` but for checkpoints both total supplies.

**Result:**

We'll have a library that `LlamaPolicy` can use to checkpoint supplies.

I recommend reviewing by:
- Diffing the `SupplyCheckpoints.sol` file against `Checkpoints.sol`
- Diffing the `SupplyCheckpoints.t.sol` file against `Checkpoints.t.sol`