feat: checkpoint role supplies (#414)
**Motivation:**

Checkpoints role supplies, which:
- Removes the need for hacks like
`_assertNoActionCreationsAtCurrentTimestamp`.
- Resolves https://github.com/llamaxyz/llama/issues/393.
- Allows strategies to use the supply at the prior block to eliminate
race conditions when role balances are updated in the same block as an
action creation.

**Modifications:**

- Removes `_assertNoActionCreationsAtCurrentTimestamp`
- Updates strategies to use the new `policy.getPastRoleSupply*` methods
- Updates tests accordingly. Notable, we add a `mineBlock()` function to
advance the block number and timestamp by 1. This is needed because we
now use historical lookups, which requires the block to be mined.

**Result:**

The above items in the motivation section are resolved.

---

Gas Diff:
- With via-ir: https://www.diffchecker.com/kQzuOofn/
- Without via-ir: https://www.diffchecker.com/KYtIHlOT/

via-ir, summary of average and median values for two relevant functions:
- `createAction`: 192,967 to 196,788 and 218,422 to 219,938
- `setRoleHolder`: 142,130 to 145,427 and 144,915 to 146,624

without via-ir, summary of average and median values for two relevant
functions:
- `createAction`:  206,182 to 211,670 and 228,203 to 233,140
- `setRoleHolder`: 141,964 to 149,075 and 148,769 to 151,257

I included both via-ir and regular results because with via-ir a few
tests fail due to how solidity handles `block.timestamp` changes in
tests, so wanted to make sure that didn't result in any big differences.

Note that the worst case increase comes in setRoleHolder, which can be
~40k gas more expensive in this checkpoint version when we have to write
new balance checkpoints AND new supply checkpoints (two new slots at 20k
each). I'm surprised the average and median difference were not nearly
that big.

I encourage you to read the gas report to make sure I have no typos
above and to check if there's any large increases I may have missed.

---

One remaining quirk is that when creating an action, the
`policy.hasPermissionId(address policyholder, uint8 role, bytes32
permissionId)` call checks if the creator has the permission ID. This
call checks:
1. If the creator holds a nonzero quantity of `role` at the current
timestamp.
2. If that role had the required permission ID at the current timestamp.

In `LlamaCore._createAction` you can find comments explaining why we're
ok with this tradeoff:

```solidity
// Typically (such as in Governor contracts) this should check that the caller has permission
// at `block.number|timestamp - 1` but here we're just checking if the caller *currently* has
// permission. Technically this introduces a race condition if e.g. an action to revoke a role
// from someone (or revoke a permission from a role) is ready to be executed at the same time as
// an action is created, as the order of transactions in the block then affects if action
// creation would succeed. However, we are ok with this tradeoff because it means we don't need
// to checkpoint the `canCreateAction` mapping which is simpler and cheaper, and in practice
// this race condition is unlikely to matter.
if (!policy.hasPermissionId(policyholder, role, permissionId)) revert PolicyholderDoesNotHavePermission();
```

I'm calling this out because we can either:
- Continue accepting this tradeoff and leave things as-is.
- Update check (1) to check if they held the role at `block.timestamp -
1`, instead of the current timestamp. (We always could have done this,
but we didn't for the reason above). Doing this makes checks (1) and (2)
more inconsistent because (1) would check the prior timestamp while (2)
would check the current timestamp. Even though it's inconsistent, it
still makes the race condition less likely to occur.
- Checkpoint the `canCreateAction` mapping so we can look things up at
the previous timestamp. We don't need full checkpointing, as we'd only
ever look up the previous timestamp, so we probably could come up with a
cheaper solution that only tracks the timestamp of the last toggle or
something.

Please share some thoughts on this remaining quirk. I'm leaning towards
leaving things as-is for simplicity and because the race condition is
not likely, but I'm open to changing it if there's a strong argument for
it. I don't think this race condition opens us up to a DoS vector but
haven't thought heavily about it yet.

On a call with @0xrajath @AustinGreen we decided:
- Move forward with this checkpointing
- Fix the issues I left in comments below
- Keep the `hasPermissionId` check as-is for consistency, that way all
permission ID checks during action creation use the current
timestamp—both parts of the same permission ID check are subject to the
same race condition
- Time permitting, we can consider updating the `canCreateAction` to
instead track the last two permission ID changes at different
timestamps, and see if the costs/code-added complexity are worth it

---------

Co-authored-by: dd0sxx <contacttheotelonis@gmail.com>
Co-authored-by: Austin Green <austinzgreen@gmail.com>