chore: c4 bot submissions, gas and analysis triage. (#428)
**Motivation:**

This PR tackles all the relevant TODOs in the c4 bot submissions, gas
and analysis triage.

**Modifications:**

* Tighter pragma ranges for `ERC721NonTransferableMinimalProxy.sol` and
`LlamaGovernanceScript.sol`
* Fixing `no-op` typo in comments
* Adding `@dev` natspec to `Enum.sol`
* `RoleAssigned` event was not being emitted for `ALL_HOLDERS_ROLE`
during policy mint/burn. Appropriate tests.
* Using private rather than public for `ALL_HOLDERS_ROLE` and
`BOOTSTRAP_ROLE` constants since they never have to be referenced
externally. And updating natspec to use `@dev` instead of `@notice` for
`internal` storage variable, such that the conventions used elsewhere in
the codebase are followed.
* Changing function visibility from `public` to `external` for certain
functions.
* Using named mappings where appropriate to provide more clarity and
readability.
* Deleting event emissions from Strategies (since they're immutable once
deployed and all data about them can be retrieved from the Strategy
related events emitted from core).
* Adding missing natspec in different locations.
* Inlining `_revokeExpiredRole()` since it's used only in
`revokeExpiredRole()`
* `ExpiredRoleRevoked` event emission along with tests. Since
`revokeExpiredRole` is a permissionless external function that can be
called by anybody, we need to emit an event with the caller information
to be consistent with the rest of the codebase and for offchain
applications to track.

**Result:**

Closes #396