docs: add documentation comparing the strategy logic contracts (#443)
**Motivation:**

This can serve as a reference for instances to compare which strategy
makes the most sense for a given permission.

**Modifications:**

Added a table with a key concepts section to our documentation
directory. I also removed unused imports in the strategy files.

**Result:**

We will have a place to refer back to for comparing strategy contracts.