chore: add comment around arg order in `PolicyCheckpoints` (#470)
**Motivation:**

Closes https://github.com/llamaxyz/llama/issues/469

Throughout `PolicyCheckpoints` we usually have the order as `(timestamp,
expiration, quantity)`. This is true in the struct ordering, and the
return params of methods like `latestCheckpoint`.

The `push` method deviates from this ordering, which is not intuitive.
There are three options to mitigate this:
1. Change the order of args in `_insert`. This is is non-invasive, but
this just moves the problem. Now, `_insert` and `push` has a different
arg order than everything else. This is nicely scoped because it leaves
the diff confined to just 2 lines in the `PolicyCheckpoints` lib
2. Change the order of args in `push`: This is more invasive because you
need to change usage in the policy contract, but it does make everything
consistent
3. Do nothing and leaving a comment indicating this inconsistency for
future devs.

Option 1 feels like a weird middle ground, and in the spirit of "write
the best code possible" we'd go with option 2. However, we'd like
minimize the diff to avoid introducing bugs since we only plan on one
more fix review, so we go with option 3

**Modifications:**

- Add aforementioned comment
- Rename a test file that was missed in
https://github.com/llamaxyz/llama/pull/430

**Result:**

See above