feat: add deployment script for absolute strategies (#490)
**Motivation:**

This PR enables instances to be configured with absolute strategies
instead of only relative strategies. This means that just by using an
absolute strategy in the strategyLogic field in the instance
configuration and the proper object format for absolute strategies in
initialStrategies, instances can deploy their instance with absolute
strategies.

**Modifications:**

- Added a strategyType flag to the `deploy-instance` command in the
justfile
- Edited `DeployLlamaInstance.s.sol` to include a `strategyType` flag
for `DeployUtils`
- Added support for absolute strategies in `DeployUtils`

**Result:**

Instances can deploy with absolute strategies.